Fix: 오늘 날짜 함수 utils로 이동
