+++ src/utils/common.ts
@@ -194,3 +194,11 @@ export const logFormData = (
     }
   }
 };
+
+export const getCurrentDateInYYYYMMDD = () => {
+  const today = new Date();
+  const year = today.getFullYear();
+  const month = today.getMonth() + 1;
+  const date = today.getDate();
+  return `${year}.${month}.${date}`;
+};
